reduce RAM requirement and DB size >5x (#12)
This is a substantial rewrite of the DB builder and query engine that reduces the database size and query engine RAM requirement over 5x. The most important changes are as follows.

DB builder output format has changed as follows:

(a) The byte order is now (snp, kmer) rather than (kmer, snp).

(b) The snp encodes additional information, such as whether the kmer is forward or reverse-complement, and the offset of the SNP within the kmer.

The 56 most-signifficant bits encode what was called snp in the previous version.

The next bit encodes forward/rc.

The 5 least-significant bits encode the SNP offset within the kmer.

(c) The endianness of kmer nucleotide encoding has changed. It is now little-endian, which improves performance on intel architectures. That means, encoding a sequence of nucleotides S with an integer N now uses the least significant bits of N to represent the initial characters of S. The byte order of N in memory matches the byte order of S in memory.

The DB builder input format has changed. The DB builder can now directly read the sckmer profile files, without requiring a shell script to process them first.

The DB builder now takes 5x less time through various tricks. This saves about an hour when building a DB on 974 species with ~7 billion kmers.

The optimized DB constructed the first time you run the query builder is 6x smaller than the original DB, and consists of 2 files:

<original_db_name>_optimized_db_snps.bin
<original_db_name>_optimized_db_kmer_index.bin

These files are constructed automatically when you provide the original db as a command line argument of the query engine. You may choose to distribute these much smaller files instead of the much larger original DB, and still provide the same command line argument (referring to a phantom file that will never be accessed if the optimized DB files above are found).

In addition, the first time this program runs, it will build indexes

<original_db_name>_optimized_db_mmer_bloom_35.bin
<original_db_name>_db_lmer_index_30.bin

as appropriate for the -l and -m command line parameters. You may ship/download those indexes or allow the query engine to build them the first time it is invoked. If you choose the latter, it is best to then quit the query engine immediately after those indexes are built, and then restart it to run queries, to realize the most compact memory footprint for running queries.

It is important to note that results after all this are exactly identical to the results produced by the previous version. In particular, the bug in the previous version that forgets to output the last SNP for each input, is still present in this version. That bug will be fixed with a separate future change.

--------------------------------- detailed change log of squashed commits below -------------------

* more accurate timing for cold DB load

free memory allocated with new []

reduce memory used for read-ahead buffers in each thread

add more supported -l and -m values

add a perf tuning option to limit the fraction of DB accessed by queries
(by default the limit is 64 TB, i.e., turned off)

validated on 78GB lenient DB to not change results from previous version

* refactor loading of DB indexes

* mmap lmer_indx and mmer_present

* implement L30 M32 with compressed SNPs

* sckmer_build refactor and support within-kmer SNP offsets

* Even cleaner parsing of multi-column sckmer profiles.

Parsing speed 44% faster.

Filtering 10x faster by removing decimal -> string -> decimal conversoin,
and using integer division instead, to obtain the 6 most-signifficant
decimal digits of a number.

* parse in parallell

sort in parallel

emit output as we go, instead of buffering all of it first

overall perf 3x, results match identically old stable version

* incorporate snp offset (within kmer) as the 8 lsbs of the snp field

reverse the order of kmer and snp in the binary output
(see comment on tuples and lex/numeric sort order)

reverse order of nucleotides in binary encoding (the LSBs
of a kmer now encode the initial basepairs)

this is the first change in binary database format -- up until
this change, results matched exactly the master version

inspected output with 'hd' (hexdump) to validate against
previous version

* work in progress since 04/22/2019

* Autosave commit created on Wed May  1 15:07:24 PDT 2019.

* Autosave commit created on Thu May  2 15:32:58 PDT 2019.

* Autosave commit created on Fri May  3 15:54:43 PDT 2019.

* This is the first solid-looking run with the DB redesign.

* Autosave commit created on Mon May  6 16:34:15 PDT 2019.

* Autosave commit created on Fri May 10 13:52:48 PDT 2019.

* Fix last bug.  Validated against master (which needs huge mem) on 800 species DB.

* Fix trivial compile error on mac.

* Remove restrictions on L.

* Autosave commit created on Fri May 10 19:12:14 PDT 2019.

* Autosave commit created on Fri May 10 20:26:29 PDT 2019.